document preprocessing
