+++ docuverse/engines/preprocessors/base_preprocessor.py
@@ -1,0 +1,8 @@
+import json
+
+class BasePreprocessor:
+    def __init__(self):
+        self.queries = []
+
+    def get_queries(self):
+        return self.queries
