+++ requirements.txt
@@ -2,3 +2,4 @@ rouge-score
 pyyaml
 elasticsearch
 sentence-transformers
+python-dotenv
